DGLWrapper: fixup some issues from static analysis.
Conflicts:
	src/DGLWrapper/actions.cpp
	src/DGLWrapper/gl-auxcontext.h
	src/DGLWrapper/gl-context.cpp
	src/DGLWrapper/gl-context.h
	src/DGLWrapper/gl-objects.cpp
	src/DGLWrapper/gl-objects.h

+++ src/DGLWrapper/gl-auxcontext.cpp
@@ -575,7 +575,7 @@ void GLAuxContext::GLQueries::setupInitialState() {
 
 void GLAuxContext::GLQueries::auxDrawTexture(GLuint name, GLenum target,
                                              GLint level, GLint layer,
-                                             GLint face,
+                                             size_t face,
                                              GLenum textureBaseFormat,
                                              GLenum renderableFormat, int width,
                                              int height) {
